requirements.txt
pysqlite3
aiosqlite
python-dotenv
discord.py
sqlalchemy
.